fix(dsv4-dspark/moe): bf16 experts in the eager (COMPILE=0) path (was fp32)
The COMPILE=0 grouped-MoE ran experts in fp32 (.float()), which is (a) a
train/serve numerical gap -- the serve runs the draft MoE in bf16 -- and
(b) the bulk of the COMPILE=0 steady cost (fp32 grouped-GEMM over 256 experts,
~2.2s vs the compiled bf16 path's ~1.6s). Switch to bf16, mirroring the
validated compile path (moe_compile uses bf16 in / fp32 on-device accumulate).
Correctness is preserved by AMP option-A: fp32 master weights + optimizer keep
the cross-step accumulation precise; the matmul does not need fp32. This lets
A3 get bf16 experts + train/serve consistency at COMPILE=0 without the heavy
torch-2.12 stack. CPU fp32 oracle unchanged -> parity tests need bf16 tolerance.

+++ src/speculators/models/dsv4_dspark/backbone/moe_grouped_gemm.py
@@ -152,8 +152,17 @@ def _fused_permute_dispatch_npu(x: torch.Tensor, indices: torch.Tensor, w_flat: 
         out = moe_compile.run(w1, w3, w2, routed_input, counts,
                               experts.swiglu_limit, routed_scores.reshape(-1))    # [nb*k, dim]
     else:
-        out = swiglu_grouped(routed_input.float(), w1.float(), w3.float(), w2.float(),
-                             counts, routed_scores.reshape(-1).float(),
+        # ★ bf16 experts (was .float()=fp32). Rationale: (1) train/serve CONSISTENCY — the serve
+        # runs the draft's MoE in bf16, so fp32-train/bf16-serve is a needless numerical gap (we've
+        # been bitten by train/serve convention mismatches before); (2) SPEED — fp32 grouped-GEMM
+        # over 256 experts was the bulk of the COMPILE=0 steady cost (~2.2s vs ~1.6s); (3) CORRECTNESS
+        # is preserved by AMP option-A: fp32 MASTER weights + optimizer keep the cross-step update
+        # accumulation precise, while the matmul (bf16 in, fp32 ACCUMULATE on-device) is where bf16
+        # belongs. Mirrors the validated compile path (moe_compile._experts_grouped_mm uses bf16).
+        # NOTE: the CPU fp32 oracle (_grouped_matmul_torch) is unchanged — any eager-vs-oracle parity
+        # test must use a bf16 tolerance, not bit-exact.
+        out = swiglu_grouped(routed_input.bfloat16(), w1.bfloat16(), w3.bfloat16(), w2.bfloat16(),
+                             counts, routed_scores.reshape(-1).bfloat16(),
                              experts.swiglu_limit, _grouped_matmul)               # [nb*k, dim]
     unpermuted = torch_npu.npu_moe_token_unpermute(out.to(routed_input.dtype), sorted_idx, None)
     if nb > n:
